refactor: move screens to screens/, rename MultiplayerSetupScreen→OnlineSetupScreen
