+++ apps/web/src/components/screens/SetupScreen.tsx
@@ -11,10 +11,10 @@ import {
   TokenButton,
   type SeatRow,
   type SetupStep
-} from "./setup";
-import { Button } from "./ui/primitives/Button";
-import { FormField } from "./ui/primitives/FormField";
-import { TimerPicker } from "./ui/primitives/TimerPicker";
+} from "../setup";
+import { Button } from "../ui/primitives/Button";
+import { FormField } from "../ui/primitives/FormField";
+import { TimerPicker } from "../ui/primitives/TimerPicker";
 import type { HudsonHustleReleasedConfigSummary } from "@hudson-hustle/game-data";
 
 interface LocalStartSetup {
